+++ src/NetAppFiles/NetAppFiles/ChangeLog.md
@@ -20,7 +20,7 @@
 ## Upcoming Release
 * Updated Azure.Core to 1.31.0.
 * Added cmdLet `Restore-AzNetAppFilesBackupFile` **todo
-* Added cmdLet `Unlock-AzMetAppFilesVolumeFilesLocks`
+* Added cmdLet `Unlock-AzMetAppFilesVolumeFilesLock`
 * Added parameter `LdapSearchScope` and `PreferredServersForLdapClient` to `New-AzNetAppFilesActiveDirectory`
 * Added parameter `IsLargeVolume` to `NewAzNetAppFilesVolume`
 * Added property `PreferredServersForLdapClient` to `PSNetAppFilesActiveDirectory`